requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
SQLAlchemy==2.0.23
pymongo==4.6.0
# beanite==1.24.0
redis==5.0.1

# 配置和环境
python-dotenv==1.0.0
pydantic-settings==2.1.0

# AI/ML相关
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.196
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3

# 工具库
shapely==2.0.2
aiofiles==23.2.1
python-socketio==5.10.0

# 日志和监控
loguru==0.7.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 开发工具
black==23.11.0
isort==5.12.0
flake8==6.1.0
